requirements.txt
torch
torchvision
segmentation-models-pytorch
scikit-learn
nibabel
SimpleITK
pandas
matplotlib
Pillow
